Move CTARadioButtons to generic spot
